requirements.txt
anyio==3.7.1
boto3==1.34.7
boto3-stubs==1.34.7
botocore==1.34.7
botocore-stubs==1.34.7
certifi==2023.11.17
click==8.1.7
distro==1.8.0
dynamodm==0.0.2
fastapi==0.105.0
glob-utils==0.0.6
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-boto3-dynamodb==1.34.0
mypy-boto3-s3==1.34.0
openai==1.6.1
pinedantic
pydantic==1.10.12
Pygments==2.17.2
python-dateutil==2.8.2
python-multipart==0.0.6
rich==13.7.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
sse-starlette==1.8.2
starlette==0.27.0
tenacity==8.2.3
tqdm==4.66.1
types-awscrt==0.20.0
types-s3transfer==0.10.0
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.25.0
google-auth-oauthlib
google-auth
bs4
lxml
python-dotenv
jinja2
kubernetes